backend/requirements.txt
Django==3.2.16
psycopg2==2.9.3
python-dotenv==1.0.0
django-colorfield==0.7.2
djangorestframework==3.12.4
djoser==2.1.0
django-filter==23.5